Moved initial state and reducers to their own files
