+++ src/common/models/components.ts
@@ -1,0 +1,6 @@
+import { create_synk_delay_component } from "common/models";
+// TODO: Rename these "instruments"
+export const createInitialComponentsState = () => ({
+  synkopaterA: create_synk_delay_component("synkopaterA", 28, 28),
+  synkopaterB: create_synk_delay_component("synkopaterB", 32, 32),
+});
